requirements.txt
pygetwindow
watchdog
groq
python-dotenv